refactor: updater use StateToolTip
